pyproject.toml
[project]
name = "fastapi-saml-example"
version = "1.0.0"
description = "Getting FastAPI to work with SAML2"
authors = [
    {name = "mrunhap", email = "liubolovelife@gmail.com"},
]
dependencies = [
    "fastapi>=0.112.0",
    "pyjwt>=2.9.0",
    "pyinstaller>=4.5.1",
    "python3-saml>=1.16.0",
    "uvicorn[standard]>=0.30.1",
    "python-multipart>=0.0.9",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
